simplify SFN to use constant memory allocation
