+++ .happy/terraform/modules/batch/main.tf
@@ -11,7 +11,7 @@ resource aws_batch_job_definition batch_job_def {
   container_properties = jsonencode({
   "jobRoleArn": "${var.batch_role_arn}",
   "image": "${var.image}",
-  "memory": var.batch_container_memory_limit,
+  "memory": 8000,
   "environment": [
     {
       "name": "ARTIFACT_BUCKET",
@@ -42,11 +42,7 @@ resource aws_batch_job_definition batch_job_def {
       "value": "${var.frontend_url}"
     }
   ],
-  "vcpus": 8,
-  "linuxParameters": {
-     "maxSwap": 800000,
-     "swappiness": 60
-  },
+  "vcpus": 1,
   "logConfiguration": {
     "logDriver": "awslogs",
     "options": {
