+++ backend/layers/processing/process.py
@@ -81,7 +81,6 @@ class ProcessMain(ProcessingLogic):
             "MAX_ATTEMPTS",
             "MIGRATE",
             "REMOTE_DEV_PREFIX",
-            "TASK_TOKEN",
         ]
         env_vars = dict()
         for var in batch_environment_variables:
@@ -103,13 +102,9 @@ class ProcessMain(ProcessingLogic):
         """
         self.logger.info(f"Processing dataset version {dataset_version_id}", extra={"step_name": step_name})
         try:
-            if step_name == "download":
-                self.process_download.process(
-                    dataset_version_id, dropbox_uri, artifact_bucket, os.environ.get("TASK_TOKEN", "")
-                )
-            elif step_name == "validate":
+            if step_name == "validate":
                 self.process_validate.process(
-                    collection_version_id, dataset_version_id, artifact_bucket, datasets_bucket
+                    collection_version_id, dataset_version_id, dropbox_uri, artifact_bucket, datasets_bucket
                 )
             elif step_name == "cxg":
                 self.process_cxg.process(dataset_version_id, artifact_bucket, cxg_bucket)
